add test module and fixed bugs
