+++ hw4/moduleSort/moduleSort.vcxproj
@@ -113,7 +113,7 @@
   <ItemDefinitionGroup Condition="'$(Configuration)|$(Platform)'=='Debug|x64'">
     <ClCompile>
       <WarningLevel>Level3</WarningLevel>
-      <SDLCheck>true</SDLCheck>
+      <SDLCheck>false</SDLCheck>
       <PreprocessorDefinitions>_DEBUG;_CONSOLE;%(PreprocessorDefinitions)</PreprocessorDefinitions>
       <ConformanceMode>true</ConformanceMode>
     </ClCompile>
@@ -141,9 +141,11 @@
   <ItemGroup>
     <ClCompile Include="moduleSort.c" />
     <ClCompile Include="myQsort.c" />
+    <ClCompile Include="qsortTests.c" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="myQsort.h" />
+    <ClInclude Include="qsortTests.h" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
   <ImportGroup Label="ExtensionTargets">
