+++ hw4/moduleSort/moduleSort.vcxproj.filters
@@ -21,9 +21,15 @@
     <ClCompile Include="myQsort.c">
       <Filter>Исходные файлы</Filter>
     </ClCompile>
+    <ClCompile Include="qsortTests.c">
+      <Filter>Исходные файлы</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="myQsort.h">
+      <Filter>Файлы заголовков</Filter>
+    </ClInclude>
+    <ClInclude Include="qsortTests.h">
       <Filter>Файлы заголовков</Filter>
     </ClInclude>
   </ItemGroup>
