+++ hw4/moduleSort/qsortTests.h
@@ -1,0 +1,9 @@
+#pragma once
+
+#include <stdbool.h>
+
+//This function checks if qsort sorts the array correctly
+bool qsortCheck(int* sortingArray, int* correctArray, int arraySize);
+
+//This function checks if test executed without errors
+bool testCorrectCase();
